Script Corrections and refactoring
