--- Assets/Scripts/Background/BackgroundController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: e4c968fbf8534984c9484c594e9c4568
-timeCreated: 1471449090
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
