+++ Assets/Scripts/General/SpriteColorController.cs
@@ -1,0 +1,28 @@
+﻿using UnityEngine;
+using System.Collections;
+using Assets.Scripts.Auxiliar.MonoEnums;
+
+public class SpriteColorController : MonoBehaviour {
+
+    public EColor color;
+    private EColor _fatherColor;
+    private SpriteRenderer _spriteRenderer;
+
+	void Start () {
+        _spriteRenderer = GetComponent<SpriteRenderer>();
+    }
+	
+	// Update is called once per frame
+	void Update () {
+        if (color == _fatherColor)
+            _spriteRenderer.enabled = true;
+        else
+            _spriteRenderer.enabled = false;
+
+	}
+
+    public void SetFatherColor(EColor fatherColor)
+    {
+        _fatherColor = fatherColor;
+    }
+}
