+++ Assets/Scripts/General/SpriteColorController.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 90600b3e11d0bf740bde7d68e0b25f9f
-timeCreated: 1471451583
+guid: 9f6de9748c877f6488ea7e90162751cc
+timeCreated: 1471451082
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
