--- Assets/Scripts/Objects/Ground.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 5a9dc800a7dad3841888bac45fb39708
-folderAsset: yes
-timeCreated: 1471451576
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
